chore: utils 파일에 있는 각 함수를 기능별로 분리해서 utils 폴더 속 파일 생성
